+++ contessa/alembic/versions/480e6618700d_change_type_to_name.py
@@ -6,25 +6,40 @@ Create Date: 2019-11-04 14:12:21.483875
 
 """
 from alembic import op
-from alembic import context
-from sqlalchemy import create_engine
 import sqlalchemy as sa
-from sqlalchemy import inspect
-
+from sqlalchemy import create_engine, inspect
 
 # revision identifiers, used by Alembic.
+from contessa.models import QualityCheck, ConsistencyCheck
+
 revision = "480e6618700d"
 down_revision = "54f8985b0ee5"
 branch_labels = None
 depends_on = None
 
-config = context.config
-url = config.get_main_option("sqlalchemy.url")
-schema = config.get_main_option("schema")
-table_prefix = "quality_check"
+config = None
 
 
-def get_quality_tables():
+def get_config():
+    global config
+    if config:
+        return config
+
+    from alembic import context
+
+    config = context.config
+
+    return config
+
+
+def get(name):
+    return get_config().get_main_option(name)
+
+
+def get_quality_tables(table_prefix):
+    url = get("sqlalchemy.url")
+    schema = get("schema")
+
     engine = create_engine(url)
     inspector = inspect(engine)
 
@@ -35,9 +50,10 @@ def get_quality_tables():
 
 
 def upgrade():
-    tables = get_quality_tables()
+    schema = get("schema")
 
-    for table in tables:
+    print("Migration Quality Check")
+    for table in get_quality_tables(QualityCheck._table_prefix):
         print(f"Migrate table {table}")
         op.alter_column(
             table,
@@ -51,9 +67,10 @@ def upgrade():
 
 
 def downgrade():
-    tables = get_quality_tables()
+    schema = get("schema")
 
-    for table in tables:
+    print("Migration Quality Check")
+    for table in get_quality_tables(QualityCheck._table_prefix):
         print(f"Migrate table {table}")
 
         op.drop_column(table, "rule_name", schema=schema)