+++ contessa/base_rules.py
@@ -1,5 +1,7 @@
 import abc
 
+from contessa.settings import TIME_FILTER_DEFAULT
+
 
 class Rule(metaclass=abc.ABCMeta):
     """
@@ -21,7 +23,9 @@ class Rule(metaclass=abc.ABCMeta):
     executor_cls = None
     description = None
 
-    def __init__(self, name, type, description, time_filter=None, condition=None):
+    def __init__(
+        self, name, type, description, time_filter=TIME_FILTER_DEFAULT, condition=None
+    ):
         self.name = name
         self.type = type
         self.time_filter = time_filter