+++ contessa/models.py
@@ -1,7 +1,6 @@
 from datetime import datetime, timedelta
 from statistics import median
-from typing import Dict, Union
-from enum import Enum
+from typing import Dict
 
 import pandas as pd
 from sqlalchemy import and_, Column, DateTime, MetaData, text, UniqueConstraint
@@ -18,21 +17,9 @@ from sqlalchemy.ext.declarative import (
     declared_attr,
 )
 
+from contessa.settings import TIME_FILTER_DEFAULT
 from contessa.base_rules import Rule
 from contessa.db import Connector
-
-
-class DataQualityDimension(Enum):
-    QUALITY = "quality"
-    CONSISTENCY = "consistency"
-
-    @classmethod
-    def from_value(cls, dimension: str):
-        for item in cls:
-            if item.value == dimension:
-                return DataQualityDimension[item.name]
-        raise ValueError("Invalid data quality dimension.")
-
 
 # default schema for results is `data_quality`, but it can be overridden by passing
 # ResultTable to runner. constructing concrete model for QualityCheck (that's abstract) will
@@ -46,11 +33,12 @@ class QualityCheck(AbstractConcreteBase, DQBase):
     """
 
     __abstract__ = True
+    _table_prefix = "quality_check"
 
     id = Column(BIGINT, primary_key=True)
-    attribute = Column(TEXT)
-    rule_name = Column(TEXT)
-    rule_type = Column(TEXT)
+    attribute = Column(TEXT, nullable=False)
+    rule_name = Column(TEXT, nullable=False)
+    rule_type = Column(TEXT, nullable=False)
     rule_description = Column(TEXT)
     total_records = Column(INTEGER)
 
@@ -63,7 +51,12 @@ class QualityCheck(AbstractConcreteBase, DQBase):
     passed_percentage = Column(DOUBLE_PRECISION)
 
     status = Column(TEXT)
-    time_filter = Column(TEXT)
+    time_filter = Column(
+        TEXT,
+        default=TIME_FILTER_DEFAULT,
+        server_default=TIME_FILTER_DEFAULT,
+        nullable=False,
+    )
     task_ts = Column(TIMESTAMP(timezone=True), nullable=False, index=True)
     created_at = Column(
         DateTime(timezone=True),
@@ -87,7 +80,7 @@ class QualityCheck(AbstractConcreteBase, DQBase):
                 "rule_type",
                 "task_ts",
                 "time_filter",
-                name=f"{cls.__tablename__}_unique_quality_check",
+                name=f"{cls.__tablename__}_unique",
             ),
         )
 
@@ -160,16 +153,22 @@ class ConsistencyCheck(AbstractConcreteBase, DQBase):
     """
 
     __abstract__ = True
+    _table_prefix = "consistency_check"
 
     id = Column(BIGINT, primary_key=True)
-    type = Column(TEXT)
-    name = Column(TEXT)
+    type = Column(TEXT, nullable=False)
+    name = Column(TEXT, nullable=False)
     description = Column(TEXT)
-    left_table = Column(TEXT)
-    right_table = Column(TEXT)
+    left_table = Column(TEXT, nullable=False)
+    right_table = Column(TEXT, nullable=False)
 
     status = Column(TEXT)
-    time_filter = Column(TEXT)
+    time_filter = Column(
+        TEXT,
+        nullable=False,
+        default=TIME_FILTER_DEFAULT,
+        server_default=TIME_FILTER_DEFAULT,
+    )
     task_ts = Column(TIMESTAMP(timezone=True), nullable=False, index=True)
     created_at = Column(
         DateTime(timezone=True),
@@ -194,7 +193,7 @@ class ConsistencyCheck(AbstractConcreteBase, DQBase):
                 "right_table",
                 "task_ts",
                 "time_filter",
-                name=f"{cls.__tablename__}_unique_consistency_check",
+                name=f"{cls.__tablename__}_unique",
             ),
         )
 
@@ -209,6 +208,7 @@ class ConsistencyCheck(AbstractConcreteBase, DQBase):
         """
         Set result to consistency check object.
         """
+        self.type = check["type"]
         self.task_ts = context["task_ts"]
         self.name = check["name"]
         self.description = check["description"]
@@ -220,7 +220,6 @@ class ConsistencyCheck(AbstractConcreteBase, DQBase):
         return f"Rule ({self.type} - {self.name} - {self.task_ts})"
 
 
-# todo - maybe create also CheckTable
 class Table:
     def __init__(self, schema_name, table_name):
         self.schema_name = schema_name
@@ -238,8 +237,10 @@ class ResultTable(Table):
     Can be overridden by `table_name`.
     """
 
-    def __init__(self, schema_name, table_name, use_prefix=True):
-        table_name = f"quality_check_{table_name}" if use_prefix else table_name
+    def __init__(self, schema_name, table_name, model_cls):
+        self.model_cls = model_cls
+        prefix = model_cls._table_prefix
+        table_name = f"{prefix}_{table_name}"
         super().__init__(schema_name, table_name)
 
     def to_camel_case(self, snake_str):
@@ -257,9 +258,7 @@ class ResultTable(Table):
         return camel_case[0].title() + camel_case[1:]
 
 
-def create_default_check_class(
-    result_table: ResultTable, data_quality_dimension: Union[DataQualityDimension, str]
-):
+def create_default_check_class(result_table: ResultTable):
     """
     This will construct type/class (not object) that will have special name that its prefixed
     with its table. It's better because sqlalchemy can yell somethings if we would use same class
@@ -274,22 +273,15 @@ def create_default_check_class(
     :return: class with dynamically created name
     """
 
-    if type(data_quality_dimension) == str:
-        data_quality_dimension = DataQualityDimension.from_value(data_quality_dimension)
-
     attributedict = {
         "__tablename__": result_table.table_name,
         "id": Column(BIGINT, primary_key=True),
         "__mapper_args__": {
-            "polymorphic_identity": f"{result_table.table_name}",
+            "polymorphic_identity": result_table.table_name,
             "concrete": True,
         },
     }
-    if data_quality_dimension == DataQualityDimension.CONSISTENCY:
-        check_class = ConsistencyCheck
-    else:
-        check_class = QualityCheck
-    cls = type(result_table.clsname, (check_class,), attributedict)
+    cls = type(result_table.clsname, (result_table.model_cls,), attributedict)
     cls.metadata.schema = result_table.schema_name
     cls.__table__.schema = result_table.schema_name
     return cls