+++ contessa/settings.py
@@ -1,0 +1,1 @@
+TIME_FILTER_DEFAULT = "not_set"